// front/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        CarritoItemResponse: {
            /** Id */
            id: number;
            producto: components["schemas"]["ProductoEnItem"];
            /** Variante Id */
            variante_id: number;
            /** Talla */
            talla: string;
            /** Cantidad */
            cantidad: number;
        };
        /** CarritoResponse */
        CarritoResponse: {
            /**
             * Session Id
             * Format: uuid
             */
            session_id: string;
            /**
             * Fecha Creacion
             * Format: date-time
             */
            fecha_creacion: string;
            /** Items */
            items: components["schemas"]["CarritoItemResponse"][];
            /** Total */
            total: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}